🧹 simplify aws discovery by moving all related code into resources/discovery.go
